requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
libusb==1.0.26b5
openant==1.2.0
packaging==23.0
Panda3D==1.10.13
Pillow==9.4.0
pkg-about==1.0.8
pyusb==1.2.1
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
zipp==3.12.0
